requirements.txt
beautifulsoup4==4.12.3
lxml==5.2.2
openai==1.40.0
httpx==0.27.2
pydantic==2.7.4
python-dotenv==1.0.1
readability-lxml==0.8.1
requests==2.32.3
tenacity==8.2.3
pytest==8.2.0
streamlit==1.37.1
python-docx==1.1.2
